Move wrapper interfaces into app layer
