+++ src/main/java/com/pyamsoft/padlock/app/service/job/PadLockFrameworkJobSchedulerService.java
@@ -20,7 +20,7 @@ import android.support.annotation.NonNull;
 import com.birbit.android.jobqueue.JobManager;
 import com.birbit.android.jobqueue.scheduling.FrameworkJobSchedulerService;
 import com.pyamsoft.padlock.PadLock;
-import com.pyamsoft.padlock.dagger.wrapper.JobSchedulerCompat;
+import com.pyamsoft.padlock.app.wrapper.JobSchedulerCompat;
 import javax.inject.Inject;
 
 public class PadLockFrameworkJobSchedulerService extends FrameworkJobSchedulerService {
