+++ src/main/java/com/pyamsoft/padlock/app/wrapper/PackageManagerWrapper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.pyamsoft.padlock.dagger.wrapper;
+package com.pyamsoft.padlock.app.wrapper;
 
 import android.content.pm.ActivityInfo;
 import android.content.pm.ApplicationInfo;
