+++ src/main/java/com/pyamsoft/padlock/dagger/lock/LockScreenInteractorImpl.java
@@ -23,10 +23,10 @@ import android.support.annotation.WorkerThread;
 import com.birbit.android.jobqueue.Job;
 import com.birbit.android.jobqueue.TagConstraint;
 import com.pyamsoft.padlock.PadLockPreferences;
-import com.pyamsoft.padlock.dagger.wrapper.PackageManagerWrapper;
+import com.pyamsoft.padlock.app.wrapper.PackageManagerWrapper;
 import com.pyamsoft.padlock.dagger.PadLockDB;
 import com.pyamsoft.padlock.dagger.job.RecheckJob;
-import com.pyamsoft.padlock.dagger.wrapper.JobSchedulerCompat;
+import com.pyamsoft.padlock.app.wrapper.JobSchedulerCompat;
 import javax.inject.Inject;
 import rx.Observable;
 import rx.functions.Func1;
