+++ src/main/java/com/pyamsoft/padlock/dagger/purge/PurgeInteractorImpl.java
@@ -18,7 +18,7 @@ package com.pyamsoft.padlock.dagger.purge;
 
 import android.support.annotation.NonNull;
 import com.pyamsoft.padlock.dagger.PadLockDB;
-import com.pyamsoft.padlock.dagger.wrapper.PackageManagerWrapper;
+import com.pyamsoft.padlock.app.wrapper.PackageManagerWrapper;
 import com.pyamsoft.padlock.model.sql.PadLockEntry;
 import java.util.List;
 import javax.inject.Inject;
