+++ src/main/java/com/pyamsoft/padlock/dagger/receiver/ReceiverModule.java
@@ -19,7 +19,7 @@ package com.pyamsoft.padlock.dagger.receiver;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import com.pyamsoft.padlock.app.receiver.ApplicationInstallReceiver;
-import com.pyamsoft.padlock.dagger.wrapper.PackageManagerWrapper;
+import com.pyamsoft.padlock.app.wrapper.PackageManagerWrapper;
 import dagger.Module;
 import dagger.Provides;
 import javax.inject.Named;
