+++ src/main/java/com/pyamsoft/padlock/dagger/service/LockServiceModule.java
@@ -19,9 +19,9 @@ package com.pyamsoft.padlock.dagger.service;
 import android.content.Context;
 import com.pyamsoft.padlock.PadLockPreferences;
 import com.pyamsoft.padlock.app.service.LockServicePresenter;
-import com.pyamsoft.padlock.dagger.wrapper.PackageManagerWrapper;
+import com.pyamsoft.padlock.app.wrapper.PackageManagerWrapper;
 import com.pyamsoft.padlock.dagger.PadLockDB;
-import com.pyamsoft.padlock.dagger.wrapper.JobSchedulerCompat;
+import com.pyamsoft.padlock.app.wrapper.JobSchedulerCompat;
 import dagger.Module;
 import dagger.Provides;
 import javax.inject.Named;
