+++ src/main/java/com/pyamsoft/padlock/dagger/wrapper/JobSchedulerCompatModule.java
@@ -17,6 +17,7 @@
 package com.pyamsoft.padlock.dagger.wrapper;
 
 import android.content.Context;
+import com.pyamsoft.padlock.app.wrapper.JobSchedulerCompat;
 import dagger.Module;
 import dagger.Provides;
 import javax.inject.Singleton;
